feature: support the use of independent files to implement the load algorithm, which is convenient for expanding different algorithms in the future.
