+++ Makefile
@@ -128,6 +128,9 @@ install: default
 	$(INSTALL) -d $(INST_LUADIR)/apisix/admin
 	$(INSTALL) apisix/admin/*.lua $(INST_LUADIR)/apisix/admin/
 
+	$(INSTALL) -d $(INST_LUADIR)/apisix/balancer
+	$(INSTALL) apisix/balancer/*.lua $(INST_LUADIR)/apisix/balancer/
+
 	$(INSTALL) -d $(INST_LUADIR)/apisix/core
 	$(INSTALL) apisix/core/*.lua $(INST_LUADIR)/apisix/core/
