+++ apisix/balancer.lua
@@ -16,14 +16,10 @@
 --
 local healthcheck
 local require     = require
-local roundrobin  = require("resty.roundrobin")
 local discovery   = require("apisix.discovery.init").discovery
-local resty_chash = require("resty.chash")
 local balancer    = require("ngx.balancer")
 local core        = require("apisix.core")
 local error       = error
-local str_char    = string.char
-local str_gsub    = string.gsub
 local pairs       = pairs
 local ipairs      = ipairs
 local tostring    = tostring
@@ -35,6 +31,10 @@ local upstreams_etcd
 
 
 local module_name = "balancer"
+local pickers = {
+    roundrobin = require("apisix.balancer.roundrobin"),
+    chash = require("apisix.balancer.chash"),
+}
 
 
 local lrucache_server_picker = core.lrucache.new({
@@ -134,72 +134,13 @@ local function fetch_healthchecker(upstream, healthcheck_parent, version)
 end
 
 
-local function fetch_chash_hash_key(ctx, upstream)
-    local key = upstream.key
-    local hash_on = upstream.hash_on or "vars"
-    local chash_key
-
-    if hash_on == "consumer" then
-        chash_key = ctx.consumer_id
-    elseif hash_on == "vars" then
-        chash_key = ctx.var[key]
-    elseif hash_on == "header" then
-        chash_key = ctx.var["http_" .. key]
-    elseif hash_on == "cookie" then
-        chash_key = ctx.var["cookie_" .. key]
-    end
-
-    if not chash_key then
-        chash_key = ctx.var["remote_addr"]
-        core.log.warn("chash_key fetch is nil, use default chash_key ",
-                      "remote_addr: ", chash_key)
-    end
-    core.log.info("upstream key: ", key)
-    core.log.info("hash_on: ", hash_on)
-    core.log.info("chash_key: ", core.json.delay_encode(chash_key))
-
-    return chash_key
-end
-
-
 local function create_server_picker(upstream, checker)
-    if upstream.type == "roundrobin" then
+    local picker = pickers[upstream.type]
+    if picker then
         local up_nodes = fetch_health_nodes(upstream, checker)
         core.log.info("upstream nodes: ", core.json.delay_encode(up_nodes))
 
-        local picker = roundrobin:new(up_nodes)
-        return {
-            upstream = upstream,
-            get = function ()
-                return picker:find()
-            end
-        }
-    end
-
-    if upstream.type == "chash" then
-        local up_nodes = fetch_health_nodes(upstream, checker)
-        core.log.info("upstream nodes: ", core.json.delay_encode(up_nodes))
-
-        local str_null = str_char(0)
-
-        local servers, nodes = {}, {}
-        for serv, weight in pairs(up_nodes) do
-            local id = str_gsub(serv, ":", str_null)
-
-            servers[id] = serv
-            nodes[id] = weight
-        end
-
-        local picker = resty_chash:new(nodes)
-        return {
-            upstream = upstream,
-            get = function (ctx)
-                local chash_key = fetch_chash_hash_key(ctx, upstream)
-                local id = picker:find(chash_key)
-                -- core.log.warn("chash id: ", id, " val: ", servers[id])
-                return servers[id]
-            end
-        }
+        return picker.new(up_nodes, upstream)
     end
 
     return nil, "invalid balancer type: " .. upstream.type, 0
